Remove ID from AutoCompleteTextView to avoid ID overlapping
If we set this ID every DropDownInput will have an AutoCompleteTextView. If someone puts two DropDownInputs on the same screen, the Android system will save the state of those text views with the same ID and when restoring them it will restore the same value on both because of the overlapping IDs.